Implemented #66
